refactor(production): isolate storyboard hash recovery
